Add Lua API to get a table with a path's components.
Each component is a table with a `type` that will be either 'addSubpath', 'addLine', 'addCurve', 'addQuadCurve', or 'closeSubpath'. The rest of the component table's key-value pairs are the same as those in the tables expected by the path instance methods corresponding to the `type` value. (Note that these are not always the same methods that were used to create the path: rects, ovals, and arcs are converted to these other primitives.)

The following snippet will produce a new path that looks the same as `path`:

```
local newPath = Path.new()
local components = path:getComponents()
for index = 1, #components do
    local component = components[index]
    newPath[component.type](newPath, component)
end
```